Generic index, edit, and form views.

+++ app/helpers/view_builder_helper.rb
@@ -11,7 +11,7 @@ module ViewBuilderHelper
       view_list + actions_bar
     end
   end
-  
+
   def ttv_actions_bar(collection, buttons)
     content_tag(:div, :class => "actions-bar wat-cf") do
       content_tag(:div, :class => "buttons") do
@@ -31,6 +31,10 @@ module ViewBuilderHelper
             html += ttv_view_link_to(:back, collection)
           when :delete
             html += ttv_view_link_to(:delete, collection)
+          when :save
+            html += content_tag(:button, t("ttv.save", :default => "Save"), :type => :submit)
+          when :cancel
+            html += ttv_view_link_to(:cancel, collection)
           else
             raise "Unknown button type in view_builder_helper#ttv_actions_bar: #{btn}"
           end
@@ -70,7 +74,9 @@ module ViewBuilderHelper
     when :delete
       link_to(t("ttv.delete", :default => "Delete"), polymorphic_path([element]), :method => :delete, :confirm => t("ttv.areyosure", :default => "Are you sure?"))
     when :back
-      link_to(t("ttv.back", :default => "Back"), polymorphic_path([element]))
+      link_to(t("ttv.back", :default => "Back"), polymorphic_path([class_symbol_p([element])]))
+    when :cancel
+      link_to(t("ttv.cancel", :default => "Cancel"), polymorphic_path([class_symbol_p([element])]))
     when :list
       link_to(t("ttv.list", :default => "List"), polymorphic_path([element]))
     when :new
@@ -105,6 +111,34 @@ module ViewBuilderHelper
     end
   end
   
+  def ttv_form_field(form, field)
+    case field
+    when "ident"
+      label = form.label :ident, t("ttv.ident", :default => "Ident"), :class => :label
+      fld = form.text_field :ident, :class => 'text_field'
+    when "display_name"
+      label = form.label :display_name, t("ttv.display_name", :default => "Display Name"), :class => :label
+      fld = form.text_field :display_name, :class => 'text_field'
+    when "asset"
+      label = "".html_safe
+      fld = form.file_field :asset
+    when "party_id"
+    when "contest_id"
+    when "position"
+    when "party"
+      parties = Party.find(:all, :order => :id)
+      label = form.label :party_id, t("activerecord.attributes.contest.party_id", :default => "Party"), :class => :label
+      fld = form.collection_select(:party_id, parties, :id, :display_name)
+    when "contest"
+      contests = Contest.find(:all, :order => :id)
+      label = form.label :contest_id, t("ttv.contest_id", :default => "Contest"), :class => :label
+      fld = form.collection_select :contest_id, contests, :id, :display_name
+    else
+      raise "view_builder_helper#ttv_form_field invalid field: #{field}"
+    end
+    label + fld
+  end
+  
   def ttv_view_list_hdr headings
     hdr_section = "".html_safe
     headings.each do 
@@ -124,7 +158,6 @@ module ViewBuilderHelper
 #
 # Take a Collection of ActiveRecord models and return the underlying class name as a :symbol. This crazy conversion
 # is needed because polymorphic_path requires the classname as a symbol in order to generate a path for the whole collection.
-#
 # e.g: coll is an array of instances of the model Asset, then class_symbol_p(coll) => :assets and class_symbol_s(coll) => :asset
 #
 # <tt>coll:</tt>Collection to be converted
